Chap 2 prob 9 (#1)
* Chapter 2 problem 9 a

Initial commit of much of the skeleton of our document

Finished part a of 2.9

* WIP

Added first part of part b and some bibliography stuff

* More general book format

ditched the UIC format and added a bibliography. We'll try to work from this moving forward

* Clean up and added subsections

* Finalized 9b & c, along with some additional work on the  document format

* Parts c and d

WIP, probably ready for a clean up

* rho to r and formatting

Changed spherical coordinate \rho to r throughout, along with a fair amount of formatting, removed many equation numbers.